test: move asymmetricMatch regression test to standalone file
The pre-existing 'should not crash under intensive usage' test in
expect-extend.test.js runs 10k iterations and can exceed its 5s timeout
on slow debug+ASAN runs. Moving the regression test to its own file
avoids that unrelated flakiness.
